**ziqi** changed all train pbs to one card training ;  2 gpu do not work. All fine tuning has add one tokenizer redownload mechnism to prevent further error
