+++ HLA_orope.pbs
@@ -1,11 +1,16 @@
 #!/bin/bash
-#PBS -N llama_train_2gpu
-#PBS -l select=1:ncpus=1:mem=64gb:ngpus=2
-#PBS -l walltime=24:00:00
+#PBS -N HLA_oRope
+#PBS -l select=1:ncpus=1:mem=64gb:ngpus=1
+#PBS -l walltime=48:00:00
 #PBS -o /rds/general/user/zz4321/home/MHA2HLA/log/${PBS_JOBID}.out
 #PBS -e /rds/general/user/zz4321/home/MHA2HLA/log/${PBS_JOBID}.err
 #PBS -j oe
 #PBS -m abe
+# ✅ 设置 HuggingFace 独立缓存目录（基于 SLURM_JOB_ID 或 PBS_JOBID）
+export HF_HOME=$TMPDIR/hf_cache_${PBS_JOBID}
+
+# 可选：显示路径确认
+echo "HuggingFace cache: $HF_HOME"
 
 # ✅ 1. **激活 Conda 环境**
 eval "$(~/miniforge3/bin/conda shell.bash hook)"
@@ -14,8 +19,8 @@ conda activate MHA2HLA
 
 # ✅ 2. **定义路径**
 RDS_PROJECTS="/rds/general/user/zz4321/home"
-MODEL_PATH="$RDS_PROJECTS/MHA2HLA"
-MODEL_DIR="$MODEL_PATH/MODEL1"
+MODEL_PATH="$RDS_PROJECTS/MHA2HLA/train/HLA_oRoPE"
+MODEL_DIR="$MODEL_PATH/model"
 LOG_PATH="$MODEL_PATH/log"
 RESULTS_PATH="$MODEL_PATH/results"
 
@@ -26,6 +31,7 @@ mkdir -p "$RESULTS_PATH"
 
 # ✅ 4. **切换到工作目录**
 cd "$MODEL_PATH"
+
 
 # ✅ 5. **运行训练脚本**
 python fine_tuning.py --model_dir "$MODEL_DIR" --epochs 50 --batch_size 4 > "$LOG_PATH/result_${PBS_JOBID}.txt" 2>&1
